+++ CHANGELOG.md
@@ -129,6 +129,8 @@ Library management
 Pragmas and options
 -------------------
 
+* New verbosity `-v debug.time:100` adds time stamps to debugging output.
+
 * Profiling options are now turned on with a new `--profile` flag instead of
   abusing the debug verbosity option. (See [#5781](https://github.com/agda/agda/issues/5731).)